requirements/local.txt
asgiref==3.5.2
Django==4.1.1
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-model-utils==4.2.0
Pillow==9.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2019.3
sqlparse==0.3.1
tzdata==2022.2
Unipath==1.1